# server/internal/app/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: SolidPing API
  description: Multi-protocol monitoring and health checking API
  version: 1.0.0
  contact:
    name: Florent Clairambault - https://github.com/fclairamb/solidping/issues
  license:
    name: AGPL-3.0
servers:
  - url: http://localhost:4000
    description: Local development server
...
paths:
...
  /api/v1/system/parameters/{key}:
    get:
      tags: [System]
      summary: Get a system parameter
      description: Returns a single system parameter by key. Secret values are masked.
      operationId: getSystemParameter
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: Parameter key
          schema:
            type: string
      responses:
        "200":
          description: System parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [System]
      summary: Set a system parameter
      description: Creates or updates a system parameter value.
      operationId: setSystemParameter
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: Parameter key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSystemParameterRequest"
      responses:
        "200":
          description: Parameter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    delete:
      tags: [System]
      summary: Delete a system parameter
      description: Soft-deletes a system parameter.
      operationId: deleteSystemParameter
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: Parameter key
          schema:
            type: string
      responses:
        "204":
          description: Parameter deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # Authentication APIs (root-level)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [title, code]
      properties:
        title:
          type: string
          description: Human-readable error description
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INTERNAL_ERROR
            - VALIDATION_ERROR
            - NOT_FOUND
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT
            - ORGANIZATION_NOT_FOUND
            - USER_NOT_FOUND
            - CHECK_NOT_FOUND
            - TOKEN_NOT_FOUND
            - NO_TOKEN
            - INVALID_TOKEN
            - INVALID_ORG
            - INVALID_CREDENTIALS
            - CHECK_HAS_ACTIVE_INCIDENTS
        detail:
          type: string
          description: Detailed error explanation
        fields:
          type: array
          description: Field-level validation errors (present for VALIDATION_ERROR)
          items:
            $ref: "#/components/schemas/ValidationErrorField"
    ValidationErrorField:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
          description: Field name that failed validation
        message:
          type: string
          description: Validation error message for this field

    # Management schemas
...
    SystemParameter:
      type: object
      properties:
        key:
          type: string
          description: Parameter key
        value:
          description: Parameter value (masked with "******" for secrets)
        secret:
          type: boolean
          description: Whether this is a secret parameter
        updatedAt:
          type: string
          format: date-time
          description: When the parameter was last updated
...
    SetSystemParameterRequest:
      type: object
      properties:
        value:
          description: The parameter value to set
        secret:
          type: boolean
          description: Whether this is a secret parameter (optional, defaults to false)
...